Added a OS formatter.
